+++ neuralteleportation/layers/mergelayers.py
@@ -30,10 +30,10 @@ class Add(nn.Module, MergeLayersMixin):
             self.next_cob = torch.ones(input2.shape[1])
 
         cob1_shape = (input2.shape[1],) + tuple([1 for _ in range(input2.dim() - 2)])
-        self.prev_cob = self.prev_cob.view(cob1_shape).float()
+        self.prev_cob = self.prev_cob.view(cob1_shape).float().type_as(input1)
 
         next_cob_shape = (input2.shape[1],) + tuple([1 for _ in range(input2.dim() - 2)])
-        self.next_cob = self.next_cob.view(next_cob_shape).float()
+        self.next_cob = self.next_cob.view(next_cob_shape).float().type_as(input1)
         # print("Add, ", self.prev_cob.flatten(), ', ', self.next_cob.flatten())
 
         return torch.add(input1, self.next_cob * input2 / self.prev_cob)
